Use the route pattern instead for span names
The ROUTE_PATTERN is added into the RequestHeader.tags map if
Play is able to find a route for it. By using this for span names
instead of the actual URI of the request, names are deduped automatically.

The names will be something like what you see when you get a 404 in dev:

- /octoparts/2/cache/invalidate/part/$partId<[^/]+>
- /admin/parts/$partId<[^/]+>/show

etc.

We could clean it up to make it look nicer, but the way these patterns are
generated is not guaranteed (an internal Play implementation detail) to remain
unchanged in the future.